fix: Eliminate false positive security warnings and resolve file path issues
## Security Warning Fixes (input_sanitizer.py)

### Fixed false positives for legitimate TASKER syntax:
- ✅ Global variable placeholders: @HOSTNAME@, @VARIABLE@ no longer flagged as "unusual characters"
- ✅ Numeric comparisons: @1_stdout@>=@MAX_COUNTER@ no longer flagged as "shell metacharacters"
- ✅ Boolean conditions: @task_success@=true properly recognized
- ✅ Logical OR operations: @task1@|@task2@ no longer flagged as "shell metacharacters"
- ✅ Descriptive text with parentheses: System(@var@) no longer flagged as "suspicious content"

### Enhanced condition pattern recognition:
- Added support for comparison operators: >=, <=, >, <, =, !=
- Added support for logical OR syntax with | operator
- Added support for boolean comparisons with true/false
- Made suspicious content detection more specific (system calls require quotes)

### Security validation still works:
- Real injection attempts like @0_stdout@; rm -rf / still properly detected
- Actual system calls like system("rm -rf /") still flagged
- Command injection patterns still caught

## File Path Resolution Fixes

### Fixed file copy warnings (task_executor_main.py):
- Added existence check before attempting task file backup copy
- Eliminates "Could not copy task file to tasks directory" warnings for non-existent files
- Clean debug messages instead of confusing warnings

### Fixed path resolution in analysis script (final_comprehensive_analysis.sh):
- Changed $test_name to $test_file to preserve directory paths
- Security tests in subdirectories now execute properly
- Fixes issue where security/buffer_overflow_*.txt files weren't found

## Sleep Parsing Fix (base_executor.py)

### Fixed decimal sleep values:
- Changed int(task.get('sleep', 0)) to float(task.get('sleep', 0)) in 3 locations
- Eliminates "invalid literal for int() with base 10: '0.5'" errors
- Supports decimal sleep values like sleep=0.5

## Regression Fix (task_executor_main.py)

### Fixed end-of-file task completion logic:
- Restored original TASKER behavior: "at least one task executed + end of file = SUCCESS"
- Tasks without explicit next=never now complete successfully instead of failing
- Maintains backward compatibility with existing task files

## Test Coverage

### Added test files:
- sleep_decimal_test.txt: Tests decimal sleep values
- regression_test.txt: Tests end-of-file completion behavior

## Impact
- ✅ Eliminates confusing false positive warnings
- ✅ Maintains robust security protection
- ✅ Improves user experience
- ✅ Fixes legitimate TASKER syntax being flagged as threats
- ✅ Resolves file path issues in comprehensive testing
- ✅ No regressions in existing functionality

+++ tasker/executors/base_executor.py
@@ -182,7 +182,7 @@ class BaseExecutor(ABC):
                     'stderr': '',
                     'success': True,
                     'skipped': False,
-                    'sleep_seconds': int(task.get('sleep', 0))
+                    'sleep_seconds': float(task.get('sleep', 0))
                 }
 
             # 8. Real execution with memory-efficient streaming
@@ -230,7 +230,7 @@ class BaseExecutor(ABC):
                     'stderr': str(e),
                     'success': False,
                     'skipped': False,
-                    'sleep_seconds': int(task.get('sleep', 0))
+                    'sleep_seconds': float(task.get('sleep', 0))
                 }
 
             # 9. Handle output splitting if configured
@@ -251,7 +251,7 @@ class BaseExecutor(ABC):
                 'stderr': processed_stderr,
                 'success': success,
                 'skipped': False,
-                'sleep_seconds': int(task.get('sleep', 0))
+                'sleep_seconds': float(task.get('sleep', 0))
             }
 
         except Exception as e:
